meの書き換え
