+++ src/_store/domain/me/index.ts
@@ -1,6 +1,5 @@
 import { defineDBModule } from '@/_store/defineDBModule'
 import { state } from './state'
-import { getters } from './getters'
 import { mutations } from './mutations'
 import { actions } from './actions'
 
@@ -8,7 +7,6 @@ export const me = defineDBModule({
   path: ['domain.me.detail'],
   namespaced: true,
   state,
-  getters,
   mutations,
   actions
 })
