+++ src/components/Main/MainView/ChannelView/use/channelState.ts
@@ -1,12 +1,11 @@
 import { computed, reactive } from 'vue'
-import _store from '@/_store'
 import store from '@/store'
 import { ChannelId } from '@/types/entity-ids'
 
 const useChannelState = (props: { channelId: ChannelId }) => {
   const state = reactive({
     stared: computed(() =>
-      _store.state.domain.me.staredChannelSet.has(props.channelId)
+      store.state.domain.me.staredChannelSet.has(props.channelId)
     ),
     forced: computed(
       () =>
