+++ src/components/Main/Navigation/NavigationContent/UsersElement.vue
@@ -15,7 +15,6 @@ import { defineComponent, PropType, computed } from 'vue'
 import { User } from '@traptitech/traq'
 import UserIcon from '@/components/UI/UserIcon.vue'
 import UsersElementUserName from './UsersElementUserName.vue'
-import _store from '@/_store'
 import store from '@/store'
 import { changeDMChannelByUsername } from '@/router/channel'
 
@@ -36,7 +35,7 @@ export default defineComponent({
       store.getters.entities.DMChannelIdByUserId(props.user.id)
     )
     const hasNotification = computed(() =>
-      _store.state.domain.me.unreadChannelsMap.has(dmChannelId.value ?? '')
+      store.state.domain.me.unreadChannelsMap.has(dmChannelId.value ?? '')
     )
 
     const onClick = () => {
