+++ src/lib/websocket/message.ts
@@ -1,17 +1,8 @@
 import apis from '@/lib/apis'
 import store from '@/store'
 import _store from '@/_store'
-import { MessageCreatedEvent, MessageReadEvent } from './events'
-import { MessageId } from '@/types/entity-ids'
+import { MessageCreatedEvent } from './events'
 import { tts } from '../tts'
-
-const isMessageForCurrentChannel = (recievedChannelId: MessageId) => {
-  const currentView = _store.state.ui.mainView.primaryView
-  return (
-    (currentView.type === 'channel' || currentView.type === 'dm') &&
-    recievedChannelId === currentView.channelId
-  )
-}
 
 export const onMessageCreated = async ({ id }: MessageCreatedEvent) => {
   const res = await apis.getMessage(id)
@@ -26,16 +17,4 @@ export const onMessageCreated = async ({ id }: MessageCreatedEvent) => {
       text: res.data.content
     })
   }
-
-  if (!isMessageForCurrentChannel(res.data.channelId)) {
-    // 未読処理
-    const myId = _store.state.domain.me.detail?.id
-    if (res.data.userId !== myId) {
-      _store.commit.domain.me.upsertUnreadChannel(res.data)
-    }
-  }
-}
-
-export const onMessageRead = ({ id }: MessageReadEvent) => {
-  _store.commit.domain.me.deleteUnreadChannel(id)
 }
