+++ src/store/domain/index.ts
@@ -5,6 +5,7 @@ import { mutations } from './mutations'
 import { actions } from './actions'
 import { listeners } from './listeners'
 import { channelTree } from './channelTree'
+import { me } from './me'
 import { messagesView } from './messagesView'
 import { stampCategory } from './stampCategory'
 
@@ -21,6 +22,7 @@ export const domain = defineModule({
   actions,
   modules: {
     channelTree,
+    me,
     messagesView,
     stampCategory
   }
