+++ src/store/domain/me/state.ts
@@ -1,0 +1,16 @@
+import { StampId, ChannelId, DMChannelId } from '@/types/entity-ids'
+import { UnreadChannel, ChannelSubscribeLevel } from '@traptitech/traq'
+
+export interface S {
+  stampHistory: Map<StampId, Date>
+  unreadChannelsMap: Map<ChannelId | DMChannelId, UnreadChannel>
+  staredChannelSet: Set<ChannelId>
+  subscriptionMap: Map<ChannelId, ChannelSubscribeLevel>
+}
+
+export const state: S = {
+  stampHistory: new Map(),
+  unreadChannelsMap: new Map(),
+  staredChannelSet: new Set(),
+  subscriptionMap: new Map()
+}
